Door works, Started on button
Door opens and closes.

Added button goes up and down when collided with BUT if something exits (say a player picks up the cube while on top of the button) the button will go up when there is sill something on top. Tried fixing with a boolean but that cause the button to not move so more testing is required to fix the bug.