Added a testsfx and audio listeners
Added a feature where the audio listener changes with which camera is being viewed from. This ended up changing a lot of the fundamental camera controls (with an idea of a for loop if need be later), namely variable name changes, ensuring all cameras and audio listeners are switched off manually before starting (unless I implement the for loop). Also turning off currentPosisiton and currentAudio when a room switch occurred fixed a few unnoticed  (thanks audio listeners?)
also added a testsfx whenever the playerCube collides with what is meant to be the ground but probably isn't don't worry about it i'll fix it later