Fixed Dropping Bug
Fixed a bug where you could drop items through solid objects by adding a collider in front of and above the drop point. On an enter (aside from Player and PickUpTag tags) "droppable" is false and will not drop the held item. On leave "droppable" is set true, allowing the item to be dropped